# Use a bitfield for `families` and optimize that a bit more