I believe the `families` check is hot enough to warrant such a micro-optimization.

```
enhancers                          fastest       │ slowest       │ median        │ mean
# before:
├─ apply_modifications             416.3 µs      │ 1.21 ms       │ 464.4 µs      │ 489.2 µs
# after:
├─ apply_modifications             376.9 µs      │ 1.803 ms      │ 412.3 µs      │ 444.9 µs
```

Looks like a ~10% improvement, so why not :-)